requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioop-lts==0.2.2
backoff==2.2.1
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
croniter==6.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
filelock==3.20.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-google-genai==4.1.3
langchain-mongodb==0.10.0
langchain-neo4j==0.7.0
langchain-openai==1.1.7
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-checkpoint-mongodb==0.3.0
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.1
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mem0ai==1.0.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
neo4j==6.0.3
neo4j-graphrag==1.12.0
networkx==3.6.1
numpy==2.4.0
ollama==0.6.1
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
portalocker==3.2.0
posthog==7.5.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydentic==0.0.1.dev3
pygame==2.6.1
Pygments==2.19.2
pymongo==4.15.5
pymongo-search-utils==0.1.0
pyparsing==3.3.1
pypdf==6.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
python-stdnum==2.2
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
rank-bm25==0.2.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rq==2.6.1
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.0
sentry-sdk==2.48.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.5
SQLAlchemy==2.0.45
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer==0.21.0
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
